+++ src/main/java/indexes/averages/AVERAGE_YEAR_COMMON.java
@@ -13,12 +13,12 @@ import model.books.YearBook;
 
 public class AVERAGE_YEAR_COMMON extends MapIndex {
 	
-	public static Map<String,Double> getIndexMap(YearBook yb){		
-		Optional<Map<String, Double>> optResult = MapIndex.alreadyComputedIndex(yb,getMyClass());
+	public static Map<String,Long> getIndexMap(YearBook yb){		
+		Optional<Map<String, Long>> optResult = MapIndex.alreadyComputedIndex(yb,getMyClass());
 		if(optResult.isPresent()) {
 			return optResult.get();
 		}else {
-			Map<String,Double> map = computeYearIndex(yb);
+			Map<String,Long> map = computeYearIndex(yb);
 			MapIndex.addComputedIndex(yb, getMyClass(), map);
 			return map;
 		}
@@ -28,8 +28,8 @@ public class AVERAGE_YEAR_COMMON extends MapIndex {
 		return AVERAGE_YEAR_COMMON.class;
 	}
 
-	private static Map<String,Double> computeYearIndex(YearBook yb) {
-		Map<String,Double> map = new HashMap<>();
+	private static Map<String,Long> computeYearIndex(YearBook yb) {
+		Map<String,Long> map = new HashMap<>();
 		
 		//double amount = 0;
 		//Vector<String> counts = yb.getSchema().getCounts(); 		
@@ -49,7 +49,7 @@ public class AVERAGE_YEAR_COMMON extends MapIndex {
 							String id = Utils.delimiterizing(Constants.AVERAGE,year,common);
 							
 							if(!map.containsKey(id)) {
-								map.put(id, 0d);
+								map.put(id, 0L);
 							}
 							
 						//}
@@ -57,9 +57,9 @@ public class AVERAGE_YEAR_COMMON extends MapIndex {
 					
 				//}
 				
-				Map<String,Double> depMap = YEAR_MONTH_COMMON.getIndexMap(yb);
+				Map<String,Long> depMap = YEAR_MONTH_COMMON.getIndexMap(yb);
 				
-				Map<String,Double> tmpAmountMap = new HashMap<>();
+				Map<String,Long> tmpAmountMap = new HashMap<>();
 				Map<String,Integer> tmpIncrementMap = new HashMap<>();
 				
 				for (String key : depMap.keySet()) {
@@ -76,7 +76,7 @@ public class AVERAGE_YEAR_COMMON extends MapIndex {
 					String id = Utils.delimiterizing(Constants.AVERAGE,year,common);
 					
 					if(!tmpAmountMap.containsKey(id)) {
-						tmpAmountMap.put(id, 0d);
+						tmpAmountMap.put(id, 0L);
 					}
 					tmpAmountMap.put(id,tmpAmountMap.get(id)+depMap.get(key));
