+++ src/main/java/indexes/balances/BALANCE_YEAR_COUNT.java
@@ -13,12 +13,12 @@ import model.books.YearBook;
 
 public class BALANCE_YEAR_COUNT extends MapIndex {
 	
-	public static Map<String,Double> getIndexMap(YearBook yb){		
-		Optional<Map<String, Double>> optResult = MapIndex.alreadyComputedIndex(yb,getMyClass());
+	public static Map<String,Long> getIndexMap(YearBook yb){		
+		Optional<Map<String, Long>> optResult = MapIndex.alreadyComputedIndex(yb,getMyClass());
 		if(optResult.isPresent()) {
 			return optResult.get();
 		}else {
-			Map<String,Double> map = computeYearIndex(yb);
+			Map<String,Long> map = computeYearIndex(yb);
 			MapIndex.addComputedIndex(yb, getMyClass(), map);
 			return map;
 		}
@@ -28,8 +28,8 @@ public class BALANCE_YEAR_COUNT extends MapIndex {
 		return BALANCE_YEAR_COUNT.class;
 	}
 
-	private static Map<String,Double> computeYearIndex(YearBook yb) {
-		Map<String,Double> map = new HashMap<>();
+	private static Map<String,Long> computeYearIndex(YearBook yb) {
+		Map<String,Long> map = new HashMap<>();
 		
 		//double amount = 0;
 		Vector<String> counts = yb.getSchema().getCounts(); 
@@ -44,7 +44,7 @@ public class BALANCE_YEAR_COUNT extends MapIndex {
 								
 				// Int with balance of month
 				if(!map.containsKey(id)) {
-					map.put(id, 0d);
+					map.put(id, 0L);
 				}
 				
 				map.put(id, map.get(id)+cmt.getFinalBalance());
@@ -71,7 +71,7 @@ public class BALANCE_YEAR_COUNT extends MapIndex {
 							
 							// Int with balance of month
 							if(!map.containsKey(id)) {
-								map.put(id, 0d);
+								map.put(id, 0L);
 							}
 							
 						}
